:truck: Move util functions
