+++ cmd/prover-inputs.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"math/big"
-	"strings"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
@@ -221,18 +220,6 @@ func AddDataDirFlag(dataDir *string, f *pflag.FlagSet) string {
 	return flagName
 }
 
-func AddLogLevelFlag(logLevel *string, f *pflag.FlagSet) string {
-	flagName := "log-level"
-	f.StringVar(logLevel, flagName, "info", "Log level (debug|info|warn|error)")
-	return flagName
-}
-
-func AddLogFormatFlag(logFormat *string, f *pflag.FlagSet) string {
-	flagName := "log-format"
-	f.StringVar(logFormat, flagName, "text", "Log format (json|text)")
-	return flagName
-}
-
 func AddBlockNumberFlag(blockNumber *string, f *pflag.FlagSet) string {
 	flagName := blockNumberFlag
 	f.StringVar(blockNumber, flagName, "", "Block number")
@@ -243,41 +230,6 @@ func AddChainIDFlag(chainID *string, f *pflag.FlagSet) string {
 	flagName := "chain-id"
 	f.StringVar(chainID, flagName, "", "Chain ID (decimal)")
 	return flagName
-}
-
-// 4. Utility functions
-func setupLogger(logLevel, logFormat string) error {
-	cfg := zap.NewProductionConfig()
-
-	// Log Level
-	switch strings.ToLower(logLevel) {
-	case "debug":
-		cfg.Level.SetLevel(zap.DebugLevel)
-	case "info":
-		cfg.Level.SetLevel(zap.InfoLevel)
-	case "warn":
-		cfg.Level.SetLevel(zap.WarnLevel)
-	case "error":
-		cfg.Level.SetLevel(zap.ErrorLevel)
-	case "":
-		// do nothing, keep default from Production
-	default:
-		return fmt.Errorf("invalid log-level %q, must be one of: debug, info, warn, error", logLevel)
-	}
-
-	// Log Format
-	if strings.EqualFold(logFormat, "text") {
-		cfg.Encoding = "console"
-	} else {
-		cfg.Encoding = "json"
-	}
-
-	logger, err := cfg.Build()
-	if err != nil {
-		return fmt.Errorf("failed to build logger: %w", err)
-	}
-	zap.ReplaceGlobals(logger)
-	return nil
 }
 
 func parseBigInt(val, flagName string) (*big.Int, error) {
